chore(RHIF-277): Refactor the /routes/InventoryTable into hybridInventory tabs (#2013)
* chore(RHIF-277): refactor /routes/InventoryTable component into hybridInventoryTabs

* chore: unchange dev.webpack

* chore: fix failing tests

* chore(RHIF-278): apply PR nitpicks, remove page header from conventional systems tab

* chore: fix hybridInventory tab change

* chore: fix table actions
